Implemented create service unit test; implemented automapper mappers; Corrected BaseDbContext for username default value if null or empty; Implemented SaveChangesAsync in UnitOfWork for EntityFramework look and feel; SaveChangesAsync removed from GenericRepository because it is made in UnitOfWork.

+++ src/NonSuckingUnitOfWork.Solution/NSUOW.Application/DTOs/VolumeDto.cs
@@ -4,7 +4,7 @@ namespace NSUOW.Application.DTOs
 {
     public partial class VolumeDto : BaseDto
     {
-        public long ServiceId { get; set; }
+        public int ServiceId { get; set; }
         public string VolumeBarCode { get; set; } = null!;
         public int VolumeNumber { get; set; }
         public decimal Weight { get; set; }
